requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==2.3.7
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
gunicorn==20.1.0